+++ src/pyiem/util.py
@@ -528,6 +528,23 @@ def exponential_backoff(func, *args, **kwargs):
     return None
 
 
+def delete_property(name, cursor=None):
+    """Delete a property from the database.
+
+    Args:
+      name (str): The name of the property to delete
+      cursor (psycopg2.cursor): Optional database cursor to use
+    """
+    if cursor is None:
+        pgconn, _cursor = get_dbconnc("mesosite")
+    else:
+        _cursor = cursor
+    _cursor.execute("DELETE from properties WHERE propname = %s", (name,))
+    if cursor is None:
+        pgconn.commit()
+        pgconn.close()
+
+
 def get_properties(cursor=None):
     """Fetch the properties set
 
@@ -545,6 +562,36 @@ def get_properties(cursor=None):
     if cursor is None:
         pgconn.close()
     return res
+
+
+def set_property(name, value, cursor=None):
+    """
+    Set a property value in the database.
+
+    Args:
+      name (str): The name of the property to set
+      value (str,datetime): The value to set
+      cursor (psycopg2.cursor): Optional database cursor to use
+    """
+    if cursor is None:
+        pgconn, _cursor = get_dbconnc("mesosite")
+    else:
+        _cursor = cursor
+    # auto convert datetime to ISO9660 string
+    if isinstance(value, datetime):
+        value = value.strftime("%Y-%m-%dT%H:%M:%SZ")
+    _cursor.execute(
+        "UPDATE properties SET propvalue = %s WHERE propname = %s",
+        (value, name),
+    )
+    if _cursor.rowcount == 0:
+        _cursor.execute(
+            "INSERT into properties (propname, propvalue) VALUES (%s, %s)",
+            (name, value),
+        )
+    if cursor is None:
+        pgconn.commit()
+        pgconn.close()
 
 
 def drct2text(drct):