+++ CHANGELOG.md
@@ -18,6 +18,7 @@ an unknown station is provided.
 
 - Add `#{state}wx` hash tags to NWS Local Storm Report social media posts.
 - Add `pyiem.iemre.reproject2iemre` helper to bring arbitrary grids onto IEMRE.
+- Add `pyiem.util.{delete,set}_property` helpers.
 - Introduce `pyiem.webutil.iemapp` decorator to do fancy things for IEM
 mod_wsgi apps and remove downstream boilerplate.